Split client and orleans signalr functions into seperate hubs
RealtimeUpdatesHubClient.cs is used for all Client->Server calls.
RealtimeUpdatesHubOrleans.cs is used for all Orleans->(Server)->Client calls.

RealtimeUpdatesHubOrleans.cs needed a reference to the hubcontext of RealtimeUpdatesHubClient.cs to be able to access it's clients. Otherwise it cannot send updates to the Godot Clients, because the Clients in RealtimeUpdatesHubOrleans.cs are the orleans (grains) clients