Fixed issue where player was not fully removed from chunk after switching chunks
I didn't call LeaveChunk() in PlayerGrain.cs but instead only removed the player by calling chunk.RemovePlayer().

So the player was still connected to the old chunk's signalR group. This meant the player received updates from the old chunk where he was not a part of anymore. And also other players didn't receive an update that the player left the chunk.

Also added the grain reference for worldChunk in de PlayerState.cs so that we don't have to track the id. This did however open up a problem where ChunkGrain can be null. This is fixed in the GetCurrentChunk which will return chunk 0 if the ChunkGrain from the PlayerState.cs is null.

Also fixed an issue where the group name in signalR did not have an integer as value. It appears that `GetPrimaryKeyString()` gives something other than `this.GetPrimaryKeyLong().ToString()` which we actually need for grains with an integer key