Better way of checking if player is already in target chunk
When calling `EnterChunk` we checked if the player's `currentChunk` was the same as the `targetChunk`. This is always the case on a fresh start because the `Initialize` calls `await EnterChunk(await GetCurrentChunk());` so `targetChunk` is the `currentChunk`. In this case the player has never been added to the chunk.

So now i've implemented a `IsPlayerInChunk` which checks if the player is actually registered in the chunk.